separate the i2c over tcp from puzzle bus libraries
