+++ client/readme.md
@@ -1,17 +1,5 @@
 # puzzle box client
 
-goal (in order of implementation):
-```
-(pbc) help
-    exit                exit pbc
-    test                send a test puzbus message
-    help                show this help
-    send <addr> <data>  [debug] send raw message
-    status              show global puzzle box state (main controller state)
-    reset               reset entire game state
-    ls                  list connected puzzle modules
-```
-
 ## Send data
 
 ```
